Pharmacy Management System based on Node Js.
The name of the project is Pharmacy Manangent System (PMS). There are two type of user in this project. One is salesman and other one is admin. Admin can add salesman. A salesman basically sell the product. Admin can add the mdicine based on category, batch no and generic name. He also can edit those if it is necessary. node.js,MySQL, JavaScript, jQuery, Bootstrap, HTML, CSS these are the stack that are used in this project.